# Performance improvement: Non-recursively drop TokenStreams with a visitor