@WalkerKnapp 2025-01-11
Sure! This is what I'm seeing with that branch on my machine:
```
tokenstream_parse: elapsed=0.563 (+/- 0.039)
syn_parse: elapsed=2.240 (+/- 0.047)
```
Seems to be within margin of error of my fork, and I'd agree that the solution you gave is quite a bit simpler :). I'll close this as obsolete in that case, but I'm glad to have been able to spark the idea!