@dtolnay 2025-01-11
Published in 1.0.93.

I am glad you are profiling syn benchmarks. I have not looked at that in years.

You might be interested to pick up https://github.com/dtolnay/proc-macro2/issues/336.